+++ utils/twitchApi.js
@@ -33,15 +33,13 @@ async function getTwichClips(twitchUsername, clientId, token) {
 
 async function getStreamMakers(twitchUsername, clientId, token) {
     const twitchUserInfo = await getTwitchUserInfo(twitchUsername, clientId, token);
-    // let url = `https://api.twitch.tv/helix/streams/markers?user_id=${twitchUserInfo.id}`;
+    let url = `https://api.twitch.tv/helix/streams/markers?user_id=${twitchUserInfo.id}`;
     // https://www.twitch.tv/videos/1237325584
-    let url = `https://api.twitch.tv/helix/streams/markers?video_id=1237325584`;
+    // let url = `https://api.twitch.tv/helix/streams/markers?video_id=1235504623`;
     const result = await twitchAPI(url, clientId, token);
     if(result.error !== undefined && result.error == 'Unauthorized') {
         console.log(`Error getting stream markers: ${result.message}`);
     }
-    // token url from varibot. sipa does not need all of these scopes. this should be cleaned up and should have it's own client id. 
-    // https://id.twitch.tv/oauth2/authorize?client_id=rq2a841j8f63fndu5lnbwzwmbzamoy&redirect_uri=https://acceptdefaults.com/twitch-oauth-token-generator/&response_type=token&scope=bits:read+channel:read:redemptions+channel:manage:redemptions+channel:moderate+chat:edit+chat:read+user:edit:broadcast+channel:edit:commercial+user:read:broadcast
     // console.log(result);
     console.log(result.data);
     if(result.data.videos !== undefined && result.data.videos.markers !== undefined) {